hw06-jpa/src/main/java/ru/otus/hw/repositories/JpaCommentRepository.java
package ru.otus.hw.repositories;

import jakarta.persistence.EntityGraph;
import jakarta.persistence.EntityManager;
import jakarta.persistence.PersistenceContext;
import jakarta.persistence.TypedQuery;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.hibernate.Hibernate;
import org.springframework.stereotype.Repository;
import ru.otus.hw.models.Comment;

import java.util.List;
import java.util.Optional;

import static org.springframework.data.jpa.repository.EntityGraph.EntityGraphType.FETCH;

@Repository
@RequiredArgsConstructor
@Slf4j
public class JpaCommentRepository implements CommentRepository {
    @PersistenceContext
    private final EntityManager em;

    @Override
    public List<Comment> findAllByBookId(long bookId) {
        String psql = "SELECT c FROM Comment c where c.book.id = :id";
        EntityGraph<?> entityGraph = em.getEntityGraph("book-entity-graph");
        TypedQuery<Comment> query = em.createQuery(psql, Comment.class);
        query.setParameter("id", bookId);
        query.setHint(FETCH.getKey(), entityGraph);

        return query.getResultList();
    }

    @Override
    public Optional<Comment> findById(long id) {
        String psql = "SELECT c FROM Comment c where c.id = :id";
        EntityGraph<?> entityGraph = em.getEntityGraph("book-entity-graph");
        TypedQuery<Comment> query = em.createQuery(psql, Comment.class);
        query.setParameter("id", id);
        query.setHint(FETCH.getKey(), entityGraph);

        return query.getResultList().isEmpty() ? Optional.empty() : Optional.of(query.getResultList().get(0));
    }

    @Override
    public Comment save(Comment comment) {
        if (comment.getId() == null) {
            em.persist(comment);
            return comment;
        }
        Comment merged = em.merge(comment);
        Hibernate.initialize(merged.getBook());
        return merged;
    }

    @Override
    public void deleteById(long id) {
        Comment comment = em.find(Comment.class, id);

        if (comment != null) {
            em.remove(comment);
        }
    }
}
